+++ .gitignore
@@ -2,3 +2,4 @@
 site/
 .coverage
 coverage.xml
+rts-env/
